Moved createCoherentLinesFromPolylines to geomHelpers
