+++ src/util/__tests__/geomHelper.test.ts
@@ -1,12 +1,12 @@
 import { LatLng } from 'leaflet';
 import constants from '~/constants/constants';
-import * as GeomHelper from '../geomHelpers';
+import * as GeomHelpers from '../geomHelpers';
 
-describe('geomHelper.createCoherentLinesFromPolylines', () => {
+describe('geomHelpers.createCoherentLinesFromPolylines', () => {
     it('Concatenates 1 line into 1 line', () => {
         const line: LatLng[] = [new LatLng(1, 1), new LatLng(1, 2), new LatLng(2, 3)];
         const expectedResult: LatLng[][] = [[new LatLng(1, 1), new LatLng(1, 2), new LatLng(2, 3)]];
-        expect(GeomHelper.createCoherentLinesFromPolylines([line])).toEqual(expectedResult);
+        expect(GeomHelpers.createCoherentLinesFromPolylines([line])).toEqual(expectedResult);
     });
 
     it('Concatenates 2 lines into 1 line', () => {
@@ -15,7 +15,9 @@ describe('geomHelper.createCoherentLinesFromPolylines', () => {
         const expectedResult: LatLng[][] = [
             [new LatLng(1, 1), new LatLng(1, 2), new LatLng(2, 2), new LatLng(2, 3)]
         ];
-        expect(GeomHelper.createCoherentLinesFromPolylines([line1, line2])).toEqual(expectedResult);
+        expect(GeomHelpers.createCoherentLinesFromPolylines([line1, line2])).toEqual(
+            expectedResult
+        );
     });
 
     it('Concatenates multiple lines into 2 lines', () => {
@@ -29,21 +31,47 @@ describe('geomHelper.createCoherentLinesFromPolylines', () => {
             [new LatLng(3, 3), new LatLng(3, 4), new LatLng(4, 5)]
         ];
         expect(
-            GeomHelper.createCoherentLinesFromPolylines([line1, line2, line3, line4, line5])
+            GeomHelpers.createCoherentLinesFromPolylines([line1, line2, line3, line4, line5])
         ).toEqual(expectedResult);
     });
 });
 
-describe('geomHelper.roundLatLng', () => {
+describe('geomHelpers.roundLatLng', () => {
     const DECIMALS_IN_GEOMETRIES = constants.DECIMALS_IN_GEOMETRIES;
 
     it(`Rounds latLng to ${DECIMALS_IN_GEOMETRIES} number of decimals`, () => {
         const lat = 1.11111111111;
         const lng = 2.222222222222;
         const coordinate = new LatLng(lat, lng);
-        const roundedCoordinate = GeomHelper.roundLatLng(coordinate);
+        const roundedCoordinate = GeomHelpers.roundLatLng(coordinate);
 
         expect(roundedCoordinate.lat.toString()).toEqual(lat.toFixed(DECIMALS_IN_GEOMETRIES));
         expect(roundedCoordinate.lng.toString()).toEqual(lng.toFixed(DECIMALS_IN_GEOMETRIES));
+    });
+});
+
+describe('geomHelpers.calculateLengthFromLatLngs', () => {
+    it('Calculates length between two points - round down', () => {
+        const positions: LatLng[] = [
+            new LatLng(60.165958, 24.9436),
+            new LatLng(60.199303, 24.940759)
+        ];
+
+        // https://gps-coordinates.org/distance-between-coordinates.php
+        // calculated distance to 3711.12 meters, rounds to 3711
+        const expectedLength = 3711;
+        expect(GeomHelpers.calculateLengthFromLatLngs(positions)).toEqual(expectedLength);
+    });
+
+    it('Calculates length between two points - round up', () => {
+        const positions: LatLng[] = [
+            new LatLng(60.17520298, 24.91285215),
+            new LatLng(60.1745068, 24.91877853)
+        ];
+
+        // https://gps-coordinates.org/distance-between-coordinates.php
+        // calculated distance to 336.77 meters, rounds to 337
+        const expectedLength = 337;
+        expect(GeomHelpers.calculateLengthFromLatLngs(positions)).toEqual(expectedLength);
     });
 });
